requirements.txt
altair==4.2.0
attrs==22.1.0
backports.zoneinfo==0.2.1
bcrypt==4.0.1
beautifulsoup4==4.11.1
blinker==1.5
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
contourpy==1.0.6
cycler==0.11.0
decorator==5.1.1
entrypoints==0.4
et-xmlfile==1.1.0
extra-streamlit-components==0.1.56
Faker==15.3.3
favicon==0.7.0
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.29
htbuilder==0.6.1
idna==3.4
importlib-metadata==5.1.0
importlib-resources==5.10.0
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
lxml==4.9.1
Markdown==3.4.1
markdownlit==0.0.5
MarkupSafe==2.1.1
matplotlib==3.6.2
more-itertools==9.0.0
mysql-connector-python==8.0.31
numpy==1.23.5
openpyxl==3.0.10
packaging==21.3
pandas==1.5.2
Pillow==9.3.0
pkgutil-resolve-name==1.3.10
plotly==5.11.0
protobuf==3.20.1
pyarrow==10.0.1
pydeck==0.8.0
Pygments==2.13.0
PyJWT==2.6.0
pymdown-extensions==9.9
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.21.0
python-http-client==3.3.7
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.1
rich==12.6.0
semver==2.13.0
sendgrid==6.9.7
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
st-annotated-text==3.0.0
st-btn-select==0.1.2
st-click-detector==0.1.3
st-pages==0.3.0
starkbank-ecdsa==2.2.0
streamlit==1.18.1
streamlit-aggrid==0.3.3
streamlit-authenticator==0.2.1
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.4
streamlit-embedcode==0.1.2
streamlit-extras==0.2.4
streamlit-faker==0.0.2
streamlit-keyup==0.2.0
streamlit-on-Hover-tabs==1.0.1
streamlit-option-menu==0.3.2
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
tenacity==8.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
typing-extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.13
validators==0.20.0
watchdog==2.1.9
zipp==3.11.0
